Extract a payload retriever interface
